Change window_event_result back to a uint8_t with the only forward declaration in fwd-event.h and fix the errors that result.

+++ common/include/event.h
@@ -48,7 +48,7 @@ enum event_type : unsigned
 	EVENT_UI_USERBOX_DRAGGED
 };
 
-enum class window_event_result : unsigned
+enum class window_event_result : uint8_t
 {
 	// Window ignored event.  Bubble up.
 	ignored,
